add antiderivative over cell boundary
